add config and launch directory
