+++ rr_sensors/config/rpi.yaml
@@ -1,0 +1,1 @@
+pin: "AIO1_PIN_11"
