+++ rr_sensors/launch/analog_sensor.launch
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<launch>
+    <node pkg="rr_sensors" type="analog_sensor" name="lr_tb5000c0" output="screen"/>
+	<rosparam file="$(find rr_sensors)/config/lr_tb5000c.yaml" />
+</launch>
